+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Features
 
   * Added `suggestedMethod` method to `Route` class.
-  * Added `buttonTo`, `formTag` and `submitButtonTag` helpers.
+  * Added `buttonTo`, `formFor` and `submitButtonTag` helpers.
   * Added support for `urlFor` options to `linkTo` helper.
 
 ## v1.1.0 (2022-06-24)